feat: bluetooth full support JSON RPC send and receive
